# src/n8n/server/openapi.yml
openapi: 3.0.0
info:
  title: r8r Public API
  description: n8n-compatible public API served by r8r. Authenticate with the X-N8N-API-KEY header.
  version: 1.1.1
servers:
  - url: /api/v1
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-N8N-API-KEY
paths:
...
  /workflows/{id}:
    get: {summary: Retrieves a workflow, tags: [Workflow], responses: {'200': {description: Operation successful.}, '404': {description: The specified resource was not found.}}}
    put: {summary: Update a workflow, tags: [Workflow], responses: {'200': {description: Workflow object}}}
    delete: {summary: Delete a workflow, tags: [Workflow], responses: {'200': {description: Operation successful.}}}
...